19 将【配置。。。】项里设置的参数保存到config文件中

+++ configuration.h
@@ -72,7 +72,7 @@ void buildCenWidgetFront();
     void writeCon();
     void readCon();
     static int bright;
-    static int contrast;
+    static float contrast;
     static int defaultAngle;
 
 //    int bright;
